Deployments/Backend/requirements.txt
pandas
numpy
nltk==3.7
scikit-learn==1.1.3
Flask
tensorflow-cpu
gunicorn